requirements.txt
Flask==2.2.2
Flask-MySQLdb==2.0.0
mysql-connector-python==8.0.31
requests==2.28.2
Werkzeug==2.2.2